Split Story Summary class into it's own module
Let's not create any circular import problems
