+++ esak/comic.py
@@ -9,7 +9,7 @@ from esak import (
     exceptions,
     prices,
     series,
-    stories,
+    story_summary,
     urls,
 )
 
@@ -53,7 +53,7 @@ class ComicSchema(Schema):
     images = fields.List(fields.Url)
     creators = fields.Nested(creator.CreatorsSchema, many=True)
     characters = fields.Nested(character.CharacterSchema, many=True)
-    stories = fields.Nested(stories.StorySummarySchema, many=True)
+    stories = fields.Nested(story_summary.StorySummarySchema, many=True)
     events = fields.Nested(events.EventsSchema, many=True)
 
     class Meta:
